+++ include/FViz/Rendering/FVizRenderer.h
@@ -4,7 +4,9 @@
 #include <FViz/Core/FVizApi.h>
 #include <FViz/Core/FVizObject.h>
 #include <FViz/Core/FVizResult.h>
+#include <FViz/Core/FVizTypes.h>
 #include <FViz/Rendering/FVizCamera.h>
+#include <FViz/Rendering/FVizScalarLegend.h>
 #include <FViz/Rendering/FVizScene.h>
 
 FVIZ_EXTERN_C_BEGIN
@@ -19,6 +21,8 @@ FVIZ_API FVizCamera* fviz_renderer_camera(FVizRenderer* renderer);
 FVIZ_API void fviz_renderer_set_background(FVizRenderer* renderer, float red, float green, float blue);
 FVIZ_API void fviz_renderer_get_background(const FVizRenderer* renderer, float* red, float* green, float* blue);
 FVIZ_API void fviz_renderer_fit_camera(FVizRenderer* renderer, float padding);
+FVIZ_API void fviz_renderer_set_scalar_legend(FVizRenderer* renderer, FVizScalarLegend* legend);
+FVIZ_API FVizScalarLegend* fviz_renderer_scalar_legend(FVizRenderer* renderer);
 
 FVIZ_EXTERN_C_END