+++ internal/FViz/Rendering/FVizRendererPrivate.h
@@ -3,12 +3,14 @@
 
 #include <FViz/Core/FVizObjectPrivate.h>
 #include <FViz/Rendering/FVizRenderer.h>
+#include <FViz/Rendering/FVizScalarLegend.h>
 
 struct FVizRenderer
 {
     FVizObject base;
     FVizScene* scene;
     FVizCamera* camera;
+    FVizScalarLegend* scalar_legend;
     float background[3];
 };